+++ app/javascript/components/Pages/LendEdit.js
@@ -110,7 +110,7 @@ export default class Lend extends Component {
                     </Form>
                 </div>
 
-                { success && <Redirect to={`/lended/${this.props.book.id}`} />}
+                { success && <Redirect to={`/lent/${this.props.book.id}`} />}
 
             </div>
         )
